Fixed sizing problem

+++ src/ca/bcit/comp2522/assignments/A3/Block.java
@@ -70,8 +70,8 @@ public class Block {
      *
      * @return sizeInCm as an int
      */
-    public static int getSizeInPx() {
-        return sizeInPx;
+    public static double getSizeInPx() {
+        return (double) sizeInPx;
     }
 
     /**
